Move branch tests to base test classes
Move branch management tests from the standalone
TestIcebergBranching into BaseIcebergConnectorTest and
BaseIcebergConnectorSmokeTest so they run against all file
formats (Parquet, Avro, ORC) and all catalog backends
(JDBC, REST, Glue, Nessie, Snowflake, etc.).
